pyproject.toml
[tool.poetry]
name = "plw-pytest"
version = "0.1.0"
description = ""
authors = ["Sergei Mikurov <454004@mail.ru>"]
readme = "README.md"
packages = [{include = "plw_pytest"}]

[tool.poetry.dependencies]
python = "^3.10"
playwright = "^1.44.0"
pytest-playwright = "^0.5.0"
flake8 = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
